requirements.txt
connexion==1.1.3
keras==2.0.8
tensorflow==1.3.0
Pillow==4.0.0
numpy==1.13.1